Fix: tsファイル整理
