+++ index.html
@@ -7,7 +7,7 @@
     <title>Pixeleria Editor</title>
   </head>
   <body>
-    <pixeleria-editor dotsize="16" artwidth="64" artheight="64"></pixeleria-editor>
+    <pixeleria-editor dotsize="64" artwidth="8" artheight="8"></pixeleria-editor>
     <script type="module" src="/src/pixeleria-editor.svelte"></script>
   </body>
 </html>
